Move ContactActivity into modules
