+++ src/client/modules/Contacts/ContactActivity/state.js
